feat(module): only add the inline script if not in the DOM
* Add the inline script only if the script is not already in the DOM

Fixes #76
