+++ src/lib/api/inline.preboot.code.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {EventSelector, PrebootOptions} from '../common/preboot.interfaces';
+import {PrebootOptions} from '../common/preboot.interfaces';
 import {getNodeKeyForPreboot} from '../common/get-node-key';
 
 import {
